extract OsgiClassLoader as more general ClassLoaderFromStackOfBrooklynClassLoadingContext
so we can use and test it from XmlSerializer (ie memento-free)
