+++ core/src/main/java/org/apache/brooklyn/core/mgmt/persist/XmlMementoSerializer.java
@@ -24,9 +24,7 @@ import java.io.IOException;
 import java.io.Writer;
 import java.util.Map;
 import java.util.NoSuchElementException;
-import java.util.Stack;
 import java.util.concurrent.ExecutionException;
-import java.util.concurrent.atomic.AtomicReference;
 
 import org.apache.brooklyn.api.catalog.CatalogItem;
 import org.apache.brooklyn.api.effector.Effector;
@@ -49,9 +47,7 @@ import org.apache.brooklyn.core.effector.BasicParameterType;
 import org.apache.brooklyn.core.effector.EffectorAndBody;
 import org.apache.brooklyn.core.effector.EffectorTasks.EffectorBodyTaskFactory;
 import org.apache.brooklyn.core.effector.EffectorTasks.EffectorTaskFactory;
-import org.apache.brooklyn.core.mgmt.classloading.BrooklynClassLoadingContextSequential;
-import org.apache.brooklyn.core.mgmt.classloading.ClassLoaderFromBrooklynClassLoadingContext;
-import org.apache.brooklyn.core.mgmt.classloading.JavaBrooklynClassLoadingContext;
+import org.apache.brooklyn.core.mgmt.classloading.ClassLoaderFromStackOfBrooklynClassLoadingContext;
 import org.apache.brooklyn.core.mgmt.rebind.dto.BasicCatalogItemMemento;
 import org.apache.brooklyn.core.mgmt.rebind.dto.BasicEnricherMemento;
 import org.apache.brooklyn.core.mgmt.rebind.dto.BasicEntityMemento;
@@ -60,14 +56,12 @@ import org.apache.brooklyn.core.mgmt.rebind.dto.BasicLocationMemento;
 import org.apache.brooklyn.core.mgmt.rebind.dto.BasicManagedBundleMemento;
 import org.apache.brooklyn.core.mgmt.rebind.dto.BasicPolicyMemento;
 import org.apache.brooklyn.core.sensor.BasicAttributeSensor;
-import org.apache.brooklyn.util.core.ClassLoaderUtils;
 import org.apache.brooklyn.util.core.xstream.XmlSerializer;
 import org.apache.brooklyn.util.exceptions.Exceptions;
 import org.apache.brooklyn.util.text.Strings;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.google.common.annotations.VisibleForTesting;
 import com.thoughtworks.xstream.converters.Converter;
 import com.thoughtworks.xstream.converters.MarshallingContext;
 import com.thoughtworks.xstream.converters.SingleValueConverter;
@@ -88,7 +82,7 @@ public class XmlMementoSerializer<T> extends XmlSerializer<T> implements Memento
 
     private static final Logger LOG = LoggerFactory.getLogger(XmlMementoSerializer.class);
 
-    private final OsgiClassLoader delegatingClassLoader;
+    private final ClassLoaderFromStackOfBrooklynClassLoadingContext delegatingClassLoader;
     private LookupContext lookupContext;
     
     public XmlMementoSerializer(ClassLoader classLoader) {
@@ -97,7 +91,7 @@ public class XmlMementoSerializer<T> extends XmlSerializer<T> implements Memento
     
     public XmlMementoSerializer(ClassLoader classLoader, Map<String, String> deserializingClassRenames) {
         super(deserializingClassRenames);
-        this.delegatingClassLoader = new OsgiClassLoader(classLoader);
+        this.delegatingClassLoader = new ClassLoaderFromStackOfBrooklynClassLoadingContext(classLoader);
         xstream.setClassLoader(this.delegatingClassLoader);
         
         xstream.alias("entity", BasicEntityMemento.class);
@@ -449,7 +443,7 @@ public class XmlMementoSerializer<T> extends XmlSerializer<T> implements Memento
                     RegisteredType cat = lookupContext.lookupManagementContext().getTypeRegistry().get(catalogItemId);
                     if (cat==null) throw new NoSuchElementException("catalog item: "+catalogItemId);
                     BrooklynClassLoadingContext clcNew = CatalogUtils.newClassLoadingContext(lookupContext.lookupManagementContext(), cat);
-                    delegatingClassLoader.pushXstreamCustomClassLoader(clcNew);
+                    delegatingClassLoader.pushClassLoadingContext(clcNew);
                     customLoaderSet = true;
                 }
                 
@@ -460,7 +454,7 @@ public class XmlMementoSerializer<T> extends XmlSerializer<T> implements Memento
             } finally {
                 context.put("SpecConverter.instance", null);
                 if (customLoaderSet) {
-                    delegatingClassLoader.popXstreamCustomClassLoader();
+                    delegatingClassLoader.popClassLoadingContext();
                 }
             }
         }
@@ -481,98 +475,6 @@ public class XmlMementoSerializer<T> extends XmlSerializer<T> implements Memento
         protected void instantiateNewInstanceSettingCache(HierarchicalStreamReader reader, UnmarshallingContext context) {
             Object instance = super.instantiateNewInstance(reader, context);
             context.put("SpecConverter.instance", instance);
-        }
-    }
-    
-    @VisibleForTesting
-    static class OsgiClassLoader extends ClassLoader {
-        private final Stack<BrooklynClassLoadingContext> contexts = new Stack<BrooklynClassLoadingContext>();
-        private final Stack<ClassLoader> cls = new Stack<ClassLoader>();
-        private final AtomicReference<Thread> xstreamLockOwner = new AtomicReference<Thread>();
-        private ManagementContext mgmt;
-        private ClassLoader currentClassLoader;
-        private AtomicReference<ClassLoaderUtils> currentLoader = new AtomicReference<>();
-        private int lockCount;
-        
-        protected OsgiClassLoader(ClassLoader classLoader) {
-            setCurrentClassLoader(classLoader);
-        }
-        
-        protected void setManagementContext(ManagementContext mgmt) {
-            this.mgmt = checkNotNull(mgmt, "mgmt");
-            currentLoader.set(new ClassLoaderUtils(currentClassLoader, mgmt));
-        }
-
-        @Override
-        protected Class<?> findClass(String name) throws ClassNotFoundException {
-            return currentLoader.get().loadClass(name);
-        }
-
-        /** Must be accompanied by a corresponding {@link #popXstreamCustomClassLoader()} when finished. */
-        @SuppressWarnings("deprecation")
-        protected void pushXstreamCustomClassLoader(BrooklynClassLoadingContext clcNew) {
-            acquireXstreamLock();
-            BrooklynClassLoadingContext oldClc;
-            if (!contexts.isEmpty()) {
-                oldClc = contexts.peek();
-            } else {
-                // TODO XmlMementoSerializer should take a BCLC instead of a CL
-                oldClc = JavaBrooklynClassLoadingContext.create(mgmt, getCurrentClassLoader());
-            }
-            BrooklynClassLoadingContextSequential clcMerged = new BrooklynClassLoadingContextSequential(mgmt, oldClc, clcNew);
-            ClassLoader newCL = ClassLoaderFromBrooklynClassLoadingContext.of(clcMerged);
-            contexts.push(clcMerged);
-            cls.push(getCurrentClassLoader());
-            setCurrentClassLoader(newCL);
-        }
-
-        protected void popXstreamCustomClassLoader() {
-            synchronized (xstreamLockOwner) {
-                releaseXstreamLock();
-                setCurrentClassLoader(cls.pop());
-                contexts.pop();
-            }
-        }
-        
-        private ClassLoader getCurrentClassLoader() {
-            return currentClassLoader;
-        }
-        
-        private void setCurrentClassLoader(ClassLoader classLoader) {
-            currentClassLoader = checkNotNull(classLoader);
-            currentLoader.set(new ClassLoaderUtils(currentClassLoader, mgmt));
-        }
-        
-        protected void acquireXstreamLock() {
-            synchronized (xstreamLockOwner) {
-                while (true) {
-                    if (xstreamLockOwner.compareAndSet(null, Thread.currentThread()) || 
-                        Thread.currentThread().equals( xstreamLockOwner.get() )) {
-                        break;
-                    }
-                    try {
-                        xstreamLockOwner.wait(1000);
-                    } catch (InterruptedException e) {
-                        throw Exceptions.propagate(e);
-                    }
-                }
-                lockCount++;
-            }
-        }
-
-        protected void releaseXstreamLock() {
-            synchronized (xstreamLockOwner) {
-                if (lockCount<=0) {
-                    throw new IllegalStateException("xstream not locked");
-                }
-                if (--lockCount == 0) {
-                    if (!xstreamLockOwner.compareAndSet(Thread.currentThread(), null)) {
-                        Thread oldOwner = xstreamLockOwner.getAndSet(null);
-                        throw new IllegalStateException("xstream was locked by "+oldOwner+" but unlock attempt by "+Thread.currentThread());
-                    }
-                    xstreamLockOwner.notifyAll();
-                }
-            }
         }
     }
 }
